+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.51.3
 * Added `internal/xstring.{FromBytes([]byte),ToBytes(string)` for increase performance on `string` from/to `[]byte` conversion
 
 ## v3.51.2